requirements.txt
# Core Web Frameworks
fastapi==0.104.1
uvicorn==0.23.2

# Data Handling and Scientific Computing
pandas==2.1.2
numpy==1.26.1
yfinance==0.2.31

# Async HTTP Clients
httpx==0.25.0

# Data Validation
pydantic==2.4.2

# Environment and Configuration
python-dotenv==1.0.0

# Typing Support
typing-extensions==4.8.0

# Web Scraping and Data Collection
requests==2.31.0

# Vector Store and Embedding
faiss-cpu==1.7.4
openai==1.2.3

# Machine Learning and NLP
scikit-learn==1.3.2
transformers==4.35.2

# Microservices and Async Support
starlette==0.27.0

# Additional Utilities
langchain==0.0.335
# Web Framework
fastapi>=0.101.0
uvicorn>=0.23.2
streamlit>=1.26.0
requests>=2.31.0

# Data Processing
pandas>=2.0.3
numpy>=1.24.3
pytz>=2023.3

# Financial Data
yfinance>=0.2.28
alpha_vantage>=2.3.1

# Vector Stores & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# LLM & Agents
langchain>=0.0.267
langchain-openai>=0.0.5
crewai>=0.28.0
langgraph>=0.0.17

# Voice Processing
torch>=2.0.1
torchaudio>=2.0.2
faster-whisper>=0.9.0
gtts>=2.3.2
pydub>=0.25.1

# Data Storage & Access
pymongo>=4.5.0
httpx>=0.24.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.3.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
# Web Scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1

# Data Processing
pandas==2.2.1
numpy==1.26.3

# Financial Data
yfinance==0.2.36

# Configuration and Environment
python-dotenv==1.0.0

# Logging
logging==0.5.1.2

# Streamlit for UI
streamlit==1.33.0

# API and Microservices
fastapi==0.109.2
uvicorn==0.27.1

# Natural Language Processing (Optional)
nltk==3.8.1

# Optional: Machine Learning for Advanced Sentiment Analysis
scikit-learn==1.3.2

# Optional: For embedding and vector store
faiss-cpu==1.7.4
openai==1.13.3
tiktoken==0.6.0

# Voice Processing (Optional)
openai-whisper==20231117
gTTS==2.5.0
# Core frameworks
streamlit==1.36.0
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.1

# Agent frameworks
langchain==0.1.9
langchain-openai==0.0.5
crewai==0.28.0
langgraph==0.1.18

# Vector databases
faiss-cpu==1.7.4
pinecone-client==3.0.1

# LLM clients
openai==1.12.0
anthropic==0.8.0

# Data sources
yfinance==0.2.35
alpha_vantage==2.3.1
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.18.1
requests==2.31.0

# Data processing
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1

# Voice capabilities
openai-whisper==20231117
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.14

# Web tools
httpx==0.26.0
aiohttp==3.9.3

# Utilities
tqdm==4.66.2
loguru==0.7.2
fastapi==0.105.0
uvicorn==0.24.0
pandas==2.1.3
yfinance==0.2.32
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
SpeechRecognition==3.10.0
pyttsx3==2.90
openai-whisper==20231117
numpy==1.26.2
soundfile==0.12.1
langchain==0.0.338
langgraph==0.0.20
faiss-cpu==1.7.4
openai==1.3.7
streamlit==1.28.2
httpx==0.25.2
python-dotenv==1.0.0
crewai==0.28.0